requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
idna==3.2
Pillow==8.3.1
playsound==1.3.0
pycparser==2.20
requests==2.26.0
selenium==3.141.0
urllib3==1.26.6
